requirements.txt
PySide6==6.8.3
Pillow>=11.2.1
numpy
windows-capture>=2.0.0
av>=12.0.0
moviepy>=2.0.0
